XRDDEV-575 use error object generated from API, rename ErrorObject -> ErrorInfo in API, do not serialize nulls with jackson
