+++ src/proxy-ui-api/src/main/resources/common-application.properties
@@ -20,3 +20,5 @@ server.servlet.session.cookie.secure=true
 
 spring.servlet.multipart.max-file-size=50MB
 spring.servlet.multipart.max-request-size=50MB
+
+spring.jackson.default-property-inclusion=NON_NULL
